+++ radicle-cli/Cargo.toml
@@ -47,6 +47,10 @@ path = "../radicle-cob"
 version = "0"
 path = "../radicle-crypto"
 
+[dependencies.radicle-term]
+version = "0"
+path = "../radicle-term"
+
 [dev-dependencies]
 pretty_assertions = { version = "1.3.0" }
 radicle = { path = "../radicle", features = ["test"] }
