+++ radicle-cli/src/commands/issue.rs
@@ -218,7 +218,7 @@ pub fn run(options: Options, ctx: impl term::Context) -> anyhow::Result<()> {
         }
         Operation::React { id, reaction } => {
             if let Ok(mut issue) = issues.get_mut(&id) {
-                let (comment_id, _) = term::comment_select(&issue).unwrap();
+                let (comment_id, _) = term::io::comment_select(&issue).unwrap();
                 issue.react(*comment_id, reaction, &signer)?;
             }
         }
