+++ radicle-term/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "radicle-term"
+license = "MIT OR Apache-2.0"
+version = "0.1.0"
+authors = ["Alexis Sellier <alexis@radicle.xyz>"]
+edition = "2021"
+
+[dependencies]
+anyhow = { version = "1" }
+concolor = { version = "0", features = ["api_unstable"] }
+inquire = { version = "0.5.3", default-features = false, features = ["termion", "editor"] }
+once_cell = { version = "1.13" }
+termion = { version = "2" }
+unicode-width = { version = "0.1.10", default-features = false }
+unicode-segmentation = { version = "1.7.1" }
+zeroize = { version = "1.1" }
+
+[dev-dependencies]
+pretty_assertions = { version = "1.3.0" }
+tempfile = { version = "3.3.0" }
