+++ radicle-term/src/ansi/paint.rs
